refactor(rpc): add maybe_async to RawTransactionRpc and centralize types
Enable async/sync flexibility for RawTransactionRpc trait with maybe_async.
Remove corepc_types dependency from floresta-node, centralizing all RPC
types in floresta-rpc as single source of truth.
