+++ crates/floresta-node/Cargo.toml
@@ -11,7 +11,6 @@ authors = ["Floresta Developers <contact@getfloresta.org>"]
 # All dependencies MUST be pinned precisely with `X.Y.z`
 axum = { workspace = true, optional = true }
 bitcoin = { workspace = true }
-corepc-types = { workspace = true }
 dns-lookup = { workspace = true }
 miniscript = { workspace = true, features = ["std"] }
 rcgen = { workspace = true }
