+++ crates/floresta-node/src/json_rpc/mod.rs
@@ -17,4 +17,5 @@ pub mod server;
 mod blockchain;
 mod control;
 mod network;
+mod raw_transaction;
 mod wallet;
